src/generics/tutorial.go
package generics

// ----- non-generics functions ------

// SumInts adds together the values of m.
func SumInts(m map[string]int64) int64 {
	var s int64
	for _, v := range m {
		s += v
	}
	return s
}

// SumFloats adds together the values of m.
func SumFloats(m map[string]float64) float64 {
	var s float64
	for _, v := range m {
		s += v
	}
	return s
}

// ------ generic functions that uses `type arguments` -----

// SumIntsOrFloats sums the values of map m. It supports both int64 and float64
// as types for map values.
func SumIntsOrFloats[K comparable, V int64 | float64](m map[K]V) V {
	var s V
	for _, v := range m {
		s += v
	}
	return s
}

// ----- declare type constraint -----

// Number type constraint
// NOTE: type argument で指定している `K comparable` の `comparable` も Go で事前定義されたの type constraint の一つ (the comparable constraint is predeclared in Go)
// 事前定義された type constraint は現在のところ以下2つのみでシンプルになっている
// * comparable
// * any
type Number interface {
	int64 | float64
}

// SumNumbers sums the values of map m. Its supports both integers
// and floats as map values.
func SumNumbers[K comparable, V Number](m map[K]V) V {
	var s V
	for _, v := range m {
		s += v
	}
	return s
}
